requirements.txt
arrow==0.12.1
django-cors-headers
Django==1.11.17
django-elasticsearch-dsl==0.5.1
djangorestframework==3.9.0
elasticsearch==6.1.1
elasticsearch-dsl==6.1.0
ipaddress==1.0.22
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
roman==3.1
six==1.12.0
urllib3==1.24.1